requirements.txt
pandas==2.3.3
requests==2.32.5
yfinance==0.2.66
sqlalchemy==2.0.44
psycopg2-binary==2.9.11
python-dotenv==1.2.1
pyyaml==6.0.3
fastapi[standard]
testcontainers[postgres]==4.14.1
pytest==8.4.2